It was found that sometimes R.NET reads Date value as a real number(with point). example 155579890.07, we must drop this .07 fraction part
Another modification for date is we turned off the flag that was
responcible for conversion to location time.

INFO:
When R.NET reads Date from R it reads as number of secs since (1970,1,1).
In C# we convert these seconds back to date and we can also convert it to
local time at the same time.