Few fixes done to 'Load dataset from R package' window: -if user types in a package name that is not avaialable on user's system, then we show a proper message.
-if the R package is not installed and user starts typing the name
in the first combobox, then the combobox will start filtering the
names from its list based on each character that user is typing.
For each hit, the first R package name in the filtered list is
selected and the datasets names related to this selected R package name
will be pupulated in the second combobox. As soon as user types in
a character that does not create a match anymore then the second
combobox was still populated with the dataset names from the last
matched package name.
This has been fixed and now in this case combobox becomes empty, as
the matching R package name is not found in the first combobox.