UTCoffset adjustment for POSIXct column has been added on C# side for making R and C# values appear same.
Also, UTCOffset property has been added that will be shown in the vargrid.
This will hold the UTC offset value, if the column is of POSIXct class. For
other columns (non-POSIXct) 0.0 will appear that can be ignored as it does
not make sense to have UTC offset for say 'gender'.