(1) If a non active dataset(which is in loaded in the datagrid but is not visble in front) is reloaded, reopened or refreshed we now activate it and show its datagrid. Basically, if right now Dataset5 is active dataset and we run some analysis or operation that refreshes Dataset1 then we switch to Dataset1 making it an active dataset (currently selected dataset).
(2) If say Dataset2 is loaded and using data manipulation command on Dataset5
user creates a new dataset with the name Dataset2 (which is already in
the datagrid) then we replace/refresh the Dataset2 but we do not drop the
associated filename because user can still go and run "Reload from disk"
to get back original Dataset2
If in above example Dataset5 happens to be an Excel file we get the
sheetname too but there may not be a sheetname related Dataset2. So
we check the filename if its not Excel file then we empty the sheetname.