Exception handling when the same key is added to QProDatasetInfoList. If the key already exists, I remove the key and then add it back. If its a new key then remove does nothing and the key gets added. Basically, we always add a key no matter if it is new or existing.
When user fills the Qpro open dataset dialog and click OK. If the dataset
name entered by an user is alreaady loaded in the grid, then
we show an overwrite prompt.
If user agrees to overwrite we first remove the existing datasetname
(which is a key) and then add this to QProDatasetInfoList.
If user do no agree the whole process is aborted.